+++ cmd/root.go
@@ -1,0 +1,16 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "ssh-hosts",
+	Short: "SSH hosts manager",
+	Long:  "Manage ~/.ssh/config",
+	Run:   ls, // 何も指定しなければlsと同じ挙動
+}
+
+func Execute() error {
+	return rootCmd.Execute()
+}
