+++ model/host.go
@@ -1,0 +1,7 @@
+package model
+
+type Host struct {
+	Num      int
+	Host     string
+	HostName string
+}
